nhi xinh

+++ HeCopUI_Framework/Controls/Clock/HClock.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Drawing;
 using System.Linq;
 using System.Text;
@@ -68,17 +69,17 @@ namespace HeCopUI_Framework.Controls.Clock
         {
             int w = this.Width;
             int h = this.Height;
-            int cx = w / 2;
-            int cy = h / 2;
-            int radius = Math.Min(cx, cy);
+            float cx = w / 2;
+            float cy = h / 2;
+            float radius = Math.Min(cx, cy);
 
             for (int i = 0; i < 60; i++)
             {
                 double angle = i * 6 * Math.PI / 180;
-                int x1 = (int)(cx + (radius - 10) * Math.Cos(angle));
-                int y1 = (int)(cy + (radius - 10) * Math.Sin(angle));
-                int x2 = (int)(cx + radius * Math.Cos(angle));
-                int y2 = (int)(cy + radius * Math.Sin(angle));
+                float x1 = (int)(cx + (radius - 10) * Math.Cos(angle));
+                float y1 = (int)(cy + (radius - 10) * Math.Sin(angle));
+                float x2 = (int)(cx + radius * Math.Cos(angle));
+                float y2 = (int)(cy + radius * Math.Sin(angle));
                 using (var penTick = new Pen(this.ticksColor, (i % 5 == 0) ? 2 : 1))
                     g.DrawLine(penTick, x1, y1, x2, y2);
 
@@ -86,8 +87,8 @@ namespace HeCopUI_Framework.Controls.Clock
                 {
                     string number = (i / 5 == 0) ? "12" : (i / 5).ToString();
                     SizeF numberSize = g.MeasureString(number, this.Font);
-                    int nx = (int)(cx + (radius - 30) * Math.Cos(angle) - numberSize.Width / 2);
-                    int ny = (int)(cy + (radius - 30) * Math.Sin(angle) - numberSize.Height / 2);
+                    float nx = (float)(cx + (radius - 30) * Math.Cos(angle) - numberSize.Width / 2);
+                    float ny = (float)(cy + (radius - 30) * Math.Sin(angle) - numberSize.Height / 2);
                     using (var brushTick = new SolidBrush(this.ticksColor))
                         g.DrawString(number, this.Font, brushTick, nx, ny);
                 }
@@ -98,9 +99,9 @@ namespace HeCopUI_Framework.Controls.Clock
         {
             int w = this.Width;
             int h = this.Height;
-            int cx = w / 2;
-            int cy = h / 2;
-            int radius = Math.Min(cx, cy);
+            float cx = w / 2;
+            float cy = h / 2;
+            float radius = Math.Min(cx, cy);
 
             DateTime now = DateTime.Now;
             DrawHandleClock(g, cx, cy, radius - 50, (now.Hour % 12) * 30 + now.Minute * 0.5, this.hrColor, 6);
@@ -108,11 +109,11 @@ namespace HeCopUI_Framework.Controls.Clock
             DrawHandleClock(g, cx, cy, radius - 20, now.Second * 6, this.secColor, 2);
         }
 
-        private void DrawHandleClock(Graphics g, int cx, int cy, int length, double angle, Color color, int width)
+        private void DrawHandleClock(Graphics g, float cx, float cy, float length, double angle, Color color, float width)
         {
             angle = angle * Math.PI / 180;
-            int x = (int)(cx + length * Math.Cos(angle - Math.PI / 2));
-            int y = (int)(cy + length * Math.Sin(angle - Math.PI / 2));
+            float x = (float)(cx + length * Math.Cos(angle - Math.PI / 2));
+            float y = (float)(cy + length * Math.Sin(angle - Math.PI / 2));
             using (var pen = new Pen(color, width))
                 g.DrawLine(pen, cx, cy, x, y);
         }
